iniital commit of Java generator
