+++ generators/java/templates/springboot/settings.gradle.ejs
@@ -1,0 +1,6 @@
+pluginManagement {
+    repositories {
+       gradlePluginPortal()
+    }
+}
+rootProject.name = '<%- lowerCaseName %>'
