+++ generators/java/templates/springboot/src/main/java/__package_dir__/UnauthorizedException.java.ejs
@@ -1,0 +1,22 @@
+package <%- basePackageName %>;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+
+@ResponseStatus(value=HttpStatus.UNAUTHORIZED)
+public class UnauthorizedException extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+
+    public UnauthorizedException() {
+        super();
+    }
+
+    public UnauthorizedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
