+++ generators/java/templates/springboot/src/main/java/__package_dir__/__class_name_prefix__Application.java.ejs
@@ -1,0 +1,12 @@
+package <%- basePackageName %>;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class <%- classNamePrefix %>Application {
+    public static void main(String... args) throws Exception {
+        SpringApplication.run(<%- classNamePrefix %>Application.class, args);
+    }
+}
