+++ generators/java/templates/springboot/src/main/java/__package_dir__/__class_name_prefix__Controller.java.ejs
@@ -1,0 +1,47 @@
+package <%- basePackageName %>;
+
+import java.util.Collections;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.smartthings.sdk.smartapp.core.SmartApp;
+import com.smartthings.sdk.smartapp.core.SmartAppDefinition;
+import com.smartthings.sdk.smartapp.core.extensions.HttpVerificationService;
+import com.smartthings.sdk.smartapp.core.models.AppLifecycle;
+import com.smartthings.sdk.smartapp.core.models.ExecutionRequest;
+import com.smartthings.sdk.smartapp.core.models.ExecutionResponse;
+
+
+@RestController
+public class <%- classNamePrefix %>Controller {
+    private final Logger log = LoggerFactory.getLogger(this.getClass());
+
+    private final SmartApp smartApp;
+    private final HttpVerificationService httpVerificationService;
+
+    public <%- classNamePrefix %>Controller(SmartAppDefinition smartAppDefinition,
+            HttpVerificationService httpVerificationService) {
+        smartApp = SmartApp.of(smartAppDefinition);
+        this.httpVerificationService = httpVerificationService;
+    }
+
+    @PostMapping("/smartapp")
+    public ExecutionResponse handle(@RequestBody ExecutionRequest executionRequest, HttpServletRequest request) {
+        log.debug("incoming request: " + executionRequest);
+        Map<String, String> headers = Collections.list(request.getHeaderNames()).stream()
+                .collect(Collectors.toMap(name -> name, name -> request.getHeader(name)));
+        if (executionRequest.getLifecycle() != AppLifecycle.PING
+                && !httpVerificationService.verify(request.getMethod(), request.getRequestURI(), headers)) {
+            throw new UnauthorizedException("unable to verify request");
+        }
+        return smartApp.execute(executionRequest);
+    }
+}
