+++ generators/java/templates/springboot/src/main/java/__package_dir__/handlers/__class_name_prefix__EventHandler.java.ejs
@@ -1,0 +1,24 @@
+package <%- basePackageName %>.handlers;
+
+import com.smartthings.sdk.smartapp.core.Response;
+import com.smartthings.sdk.smartapp.core.extensions.EventHandler;
+import com.smartthings.sdk.smartapp.core.models.EventData;
+import com.smartthings.sdk.smartapp.core.models.EventResponseData;
+import com.smartthings.sdk.smartapp.core.models.ExecutionRequest;
+import com.smartthings.sdk.smartapp.core.models.ExecutionResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+
+@Component
+public class <%- classNamePrefix %>EventHandler implements EventHandler {
+    private final Logger log = LoggerFactory.getLogger(this.getClass());
+
+    @Override
+    public ExecutionResponse handle(ExecutionRequest executionRequest) throws Exception {
+        EventData eventData = executionRequest.getEventData();
+        log.debug("EVENT: executionRequest = " + executionRequest);
+        return Response.ok(new EventResponseData());
+    }
+}
