+++ generators/java/templates/springboot/src/main/resources/application.yml.ejs
@@ -1,0 +1,4 @@
+<% if (contextStore == 'dynamodb') { %>
+smartThingsOAuthClientId: <%- clientId %>
+smartThingsOAuthClientSecret: <%- clientSecret %>
+<% } %>
