+++ generators/java/templates/springboot/src/main/resources/logback.xml.ejs
@@ -1,0 +1,15 @@
+<configuration>
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{YYYY-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+
+    <root level="INFO">
+        <appender-ref ref="STDOUT"/>
+    </root>
+
+    <logger name="com.smartthings.sdk" level="TRACE"/>
+    <logger name="<%- basePackageName %>" level="TRACE"/>
+    <logger name="app" level="TRACE"/>
+</configuration>
