+++ generators/java/templates/springboot/src/test/java/__package_dir__/__class_name_prefix__ApplicationTests.java.ejs
@@ -1,0 +1,15 @@
+package <%- basePackageName %>;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class <%- classNamePrefix %>ApplicationTests {
+    @Test
+        public void contextLoads() {
+    }
+}
