+++ generators/node/templates/app-smartapp/README.md
