+++ generators/node/templates/app-smartapp/app.js
