+++ generators/node/templates/app-smartapp/gitignore
