+++ generators/node/templates/app-smartapp/jsconfig.json
