+++ generators/node/templates/app-smartapp/package.json
