+++ generators/node/templates/app-smartapp/vscode/extensions.json
