+++ generators/node/templates/app-smartapp/vscode/launch.json
