+++ generators/node/templates/app-smartapp/vscodeignore
