+++ generators/node/validator.js
